feat: support searching/filtering label names with "l" hotkey
Resolves: #44

This is a minimal attempt at supporting searching labels found in a
scrape. It works by adding another internal boolean to identify if a
user is currently performing a label search, and then uses the existing
row filtering mechanism to fuzzy match against the string returned by
`scrape.LabelStatsSlice.String()`. This string is a concatenated list of
the labels found on a given metric and their respective cardinality
values, separated by `|`s.

Pros of this approach:
- Relatively simple code change to support label search.
- Similar speed to metric name search, since we're just comparing a
  different string in the filtering phase. If we were to store another
map or iterate over the already held `scrape.SeriesMap`, it'd slow down
filtering as we would need to then perform X number of fuzzy matches to
all label names/values. This would be especially noticeable on large
scrapes.
- Similar resource usage to previous builds, since we're (ab)using the
  existing internal structures and filtering mechanisms.

Cons of this approach:
- Because fuzzy matching is done against a built up, concatenated
  string, fuzzy matching is applied to _all_ label names at once, not
each label individually, which can potentially return false positive
matches, especially against metrics with lots of labels.
- Label search only works on label names, not label values.

I would argue this initial functionality is enough to justify adding the
feature, it can be further improved in follow up work if needed/desired.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>