feat: allow searching both metrics and labels simultaneously
Using the new enum pattern, this introduces the ability to search for
either metric names or label names in a single search. The following
user-facing changes have occurred to make this a bit more
consistent/straight-forward to use:

- metric name search has been moved from the `/` hotkey -> `m` hotkey.
  this mirrors the label name search hotkey of `l`.
- searching across both metric and label names is now on the `/` hotkey.
  the `/` key is often used for search in other software and it felt
appropriate to make it's functionality the most catch-all of searching
everything.

Not required for the given changes, but this commit also includes from
renaming of the enum values from the pattern of `$thingSearch` ->
`search$thing` so it's clear the consts are all related and it reads
more naturally.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>